Find the locally built OSMesa libs on mac/linux, and add them to the dynamic library search path.
